requirements.txt
langchain
langgraph
langchain-community
langchain-core
langchain-groq
faiss-cpu
streamlit
python-dotenv
langchain-tavily
tavily-python
pydantic